fix paralell test
