+++ app/__test__/helpers/valoratiosnOrderSetup.js
@@ -1,28 +1,28 @@
-import { encryptPassword } from "../api/v1/auth/utils";
-import { PrismaClient } from "@prisma/client";
+import { encryptPassword } from '../../api/v1/auth/utils';
+import { PrismaClient } from '@prisma/client';
 
 const prisma = new PrismaClient();
 
 export const setupV = async () => {
-  const contrasena = await encryptPassword("Contra123456");
+  const contrasena = await encryptPassword('Contra123456');
 
   const user = {
-    correo: "correo_empresa@example.com",
-    contrasena: "contrasena_segura",
-    tipo_usuario: "Empresa",
-    estatus: "Activo",
+    correo: 'correo_empresa@example.com',
+    contrasena: 'contrasena_segura',
+    tipo_usuario: 'Empresa',
+    estatus: 'Activo',
     empresa: {
-      razon_social: "Mi Empresa S.A.",
-      tipo_documento_empresa: "NIT",
-      numero_documento_empresa: "123456789",
-      telefono: "123-456-7890",
-      sitio_web: "https://www.miempresa.com",
-      camara_comercio: "Ruta a la cámara de comercio",
-      representante_legal: "Nombre del representante legal",
-      tipo_documento_representante: "Cédula",
-      numero_documento_representante: "987654321",
-      correo_representante: "representante@example.com",
-      descripcion: "Descripción de la empresa",
+      razon_social: 'Mi Empresa S.A.',
+      tipo_documento_empresa: 'NIT',
+      numero_documento_empresa: '123456789',
+      telefono: '123-456-7890',
+      sitio_web: 'https://www.miempresa.com',
+      camara_comercio: 'Ruta a la cámara de comercio',
+      representante_legal: 'Nombre del representante legal',
+      tipo_documento_representante: 'Cédula',
+      numero_documento_representante: '987654321',
+      correo_representante: 'representante@example.com',
+      descripcion: 'Descripción de la empresa',
     },
   };
 
@@ -56,15 +56,15 @@ export const setupV = async () => {
   });
 
   const userClient = {
-    correo: "correo_ser@example.com",
-    contrasena: "Contra123456",
-    tipo_usuario: "Cliente",
-    estatus: "Activo",
+    correo: 'correo_ser@example.com',
+    contrasena: 'Contra123456',
+    tipo_usuario: 'Cliente',
+    estatus: 'Activo',
     cliente: {
-      nombre_completo: "Nombre Completo",
-      tipo_documento: "Cédula",
-      numero_documento: "123456789",
-      telefono: "123-456-7890",
+      nombre_completo: 'Nombre Completo',
+      tipo_documento: 'Cédula',
+      numero_documento: '123456789',
+      telefono: '123-456-7890',
     },
   };
   const dataClient = await prisma.usuario.create({
@@ -91,7 +91,7 @@ export const setupV = async () => {
   //Creo la categoria
   const category = await prisma.categoria.create({
     data: {
-      nombre_categoria: "Pinturas",
+      nombre_categoria: 'Pinturas',
     },
   });
 
@@ -99,14 +99,14 @@ export const setupV = async () => {
     data: {
       id_empresa: empresa.id,
       id_categoria: category.id,
-      nombre: "Nombre del Producto",
-      descripcion: "Descripción del Producto",
-      ficha_tecnica: "Ficha Técnica del Producto",
+      nombre: 'Nombre del Producto',
+      descripcion: 'Descripción del Producto',
+      ficha_tecnica: 'Ficha Técnica del Producto',
       precio: 100,
       cantidad_disponible: 1,
       estatus: true,
-      tipo_entrega: "Entrega Rápida",
-      marca: "Marca del Producto",
+      tipo_entrega: 'Entrega Rápida',
+      marca: 'Marca del Producto',
       impuestos: 0.16,
     },
   });
