+++ app/__test__/testValorationsProductsCategorys/testEndPoints.spec.js
@@ -1,160 +1,161 @@
-import request from "supertest";
-import { app } from "../../index.js";
-import { resetDb } from "../helpers/reset-db.js";
+import request from 'supertest';
+import { app } from '../../index.js';
+import { resetDb } from '../helpers/reset-db.js';
 
-import { setup } from "../userCompanySetup.js";
-import { setupV } from "../valoratiosnOrderSetup.js";
+import { setup } from '../helpers/userCompanySetup.js';
+import { setupV } from '../helpers/valoratiosnOrderSetup.js';
+
 import {
   getIncorrectDataProduct,
   getPartialProduct,
   getProduct,
-} from "./fixtures/product.fixture.js";
+} from './fixtures/product.fixture.js';
 
-describe("Categorys Test", () => {
+describe('Categorys Test', () => {
   beforeEach(async () => {
     await resetDb();
   });
-  test("Create Category", async () => {
-    const result = await request(app).post("/api/v1/categorias").send({
-      nombre_categoria: "Pinturas",
+  test('Create Category', async () => {
+    const result = await request(app).post('/api/v1/categorias').send({
+      nombre_categoria: 'Pinturas',
     });
     expect(result.status).toBe(201);
   });
 
-  test("Create bad structure Category", async () => {
-    const result = await request(app).post("/api/v1/categorias").send({
-      nombre: "Pinturas",
+  test('Create bad structure Category', async () => {
+    const result = await request(app).post('/api/v1/categorias').send({
+      nombre: 'Pinturas',
     });
 
     expect(result.status).toBe(500);
   });
 
-  test("GetAll Categorys", async () => {
-    await request(app).post("/api/v1/categorias").send({
-      nombre_categoria: "Pinturas",
+  test('GetAll Categorys', async () => {
+    await request(app).post('/api/v1/categorias').send({
+      nombre_categoria: 'Pinturas',
     });
-    const response = await request(app).get("/api/v1/categorias");
+    const response = await request(app).get('/api/v1/categorias');
     expect(response.body.data.length).toBe(1);
   });
 
-  test("Get Category by id", async () => {
-    const result = await request(app).post("/api/v1/categorias").send({
-      nombre_categoria: "Pinturas",
+  test('Get Category by id', async () => {
+    const result = await request(app).post('/api/v1/categorias').send({
+      nombre_categoria: 'Pinturas',
     });
     const response = await request(app).get(
-      `/api/v1/categorias/${result.body.data.id}`
+      `/api/v1/categorias/${result.body.data.id}`,
     );
     expect(response.status).toBe(200);
   });
 
-  test("update Category", async () => {
-    const result = await request(app).post("/api/v1/categorias").send({
-      nombre_categoria: "Pinturas",
+  test('update Category', async () => {
+    const result = await request(app).post('/api/v1/categorias').send({
+      nombre_categoria: 'Pinturas',
     });
     const response = await request(app)
       .put(`/api/v1/categorias/${result.body.data.id}`)
       .send({
-        nombre_categoria: "Nueva Categoria",
+        nombre_categoria: 'Nueva Categoria',
       });
 
     expect(response.status).toBe(200);
   });
 
-  test("Delete Category", async () => {
-    const result = await request(app).post("/api/v1/categorias").send({
-      nombre_categoria: "Pinturas",
+  test('Delete Category', async () => {
+    const result = await request(app).post('/api/v1/categorias').send({
+      nombre_categoria: 'Pinturas',
     });
     const response = await request(app).delete(
-      `/api/v1/categorias/${result.body.data.id}`
+      `/api/v1/categorias/${result.body.data.id}`,
     );
     expect(response.status).toBe(204);
   });
 });
 
-describe("Products Test", () => {
+describe('Products Test', () => {
   beforeEach(async () => {
     await resetDb();
   });
 
-  test("Create Product and GetAll", async () => {
+  test('Create Product and GetAll', async () => {
     await setup();
-    const login = await request(app).post("/api/v1/auth/signin").send({
-      correo: "correo_empresa@example.com",
-      contrasena: "Contra123456",
+    const login = await request(app).post('/api/v1/auth/signin').send({
+      correo: 'correo_empresa@example.com',
+      contrasena: 'Contra123456',
     });
 
     expect(login.status).toBe(200);
     const token = login.body.meta.token;
-    const category = await request(app).post("/api/v1/categorias").send({
-      nombre_categoria: "Pinturas",
+    const category = await request(app).post('/api/v1/categorias').send({
+      nombre_categoria: 'Pinturas',
     });
     expect(category.status).toBe(201);
 
     const dataProduct = getProduct();
 
     const responseProduct = await request(app)
-      .post("/api/v1/productos")
-      .set("Authorization", `Bearer ${token}`)
+      .post('/api/v1/productos')
+      .set('Authorization', `Bearer ${token}`)
       .set(
-        "Content-Type",
-        `multipart/form-data; boundary=${dataProduct.getBoundary()}`
+        'Content-Type',
+        `multipart/form-data; boundary=${dataProduct.getBoundary()}`,
       )
       .send(dataProduct.getBuffer());
 
     expect(responseProduct.status).toBe(201);
-    const listProduct = await request(app).get("/api/v1/productos");
+    const listProduct = await request(app).get('/api/v1/productos');
     expect(listProduct.body.data.length).toBe(1);
   }, 10000);
-  test("Create Product with incorrect data", async () => {
+  test('Create Product with incorrect data', async () => {
     await setup();
-    const login = await request(app).post("/api/v1/auth/signin").send({
-      correo: "correo_empresa@example.com",
-      contrasena: "Contra123456",
+    const login = await request(app).post('/api/v1/auth/signin').send({
+      correo: 'correo_empresa@example.com',
+      contrasena: 'Contra123456',
     });
 
     expect(login.status).toBe(200);
     const token = login.body.meta.token;
-    const category = await request(app).post("/api/v1/categorias").send({
-      nombre_categoria: "Pinturas",
+    const category = await request(app).post('/api/v1/categorias').send({
+      nombre_categoria: 'Pinturas',
     });
     expect(category.status).toBe(201);
 
     const dataProduct = getIncorrectDataProduct();
 
     const responseProduct = await request(app)
-      .post("/api/v1/productos")
-      .set("Authorization", `Bearer ${token}`)
+      .post('/api/v1/productos')
+      .set('Authorization', `Bearer ${token}`)
       .set(
-        "Content-Type",
-        `multipart/form-data; boundary=${dataProduct.getBoundary()}`
+        'Content-Type',
+        `multipart/form-data; boundary=${dataProduct.getBoundary()}`,
       )
       .send(dataProduct.getBuffer());
 
     expect(responseProduct.status).toBe(400);
   }, 10000);
 
-  test("Get Product by ID", async () => {
+  test('Get Product by ID', async () => {
     await setup();
-    const login = await request(app).post("/api/v1/auth/signin").send({
-      correo: "correo_empresa@example.com",
-      contrasena: "Contra123456",
+    const login = await request(app).post('/api/v1/auth/signin').send({
+      correo: 'correo_empresa@example.com',
+      contrasena: 'Contra123456',
     });
 
     expect(login.status).toBe(200);
     const token = login.body.meta.token;
-    const category = await request(app).post("/api/v1/categorias").send({
-      nombre_categoria: "Pinturas",
+    const category = await request(app).post('/api/v1/categorias').send({
+      nombre_categoria: 'Pinturas',
     });
     expect(category.status).toBe(201);
 
     const dataProduct = getProduct();
 
     const responseProduct = await request(app)
-      .post("/api/v1/productos")
-      .set("Authorization", `Bearer ${token}`)
+      .post('/api/v1/productos')
+      .set('Authorization', `Bearer ${token}`)
       .set(
-        "Content-Type",
-        `multipart/form-data; boundary=${dataProduct.getBoundary()}`
+        'Content-Type',
+        `multipart/form-data; boundary=${dataProduct.getBoundary()}`,
       )
       .send(dataProduct.getBuffer());
 
@@ -164,69 +165,69 @@ describe("Products Test", () => {
 
     const response = await request(app)
       .get(`/api/v1/productos/${productId}`)
-      .set("Authorization", `Bearer ${token}`);
+      .set('Authorization', `Bearer ${token}`);
     expect(response.status).toBe(200);
 
     expect(response.body.data.id).toBe(productId);
   }, 10000);
 
-  test("Get Product by Incorrect ID", async () => {
+  test('Get Product by Incorrect ID', async () => {
     await setup();
-    const login = await request(app).post("/api/v1/auth/signin").send({
-      correo: "correo_empresa@example.com",
-      contrasena: "Contra123456",
+    const login = await request(app).post('/api/v1/auth/signin').send({
+      correo: 'correo_empresa@example.com',
+      contrasena: 'Contra123456',
     });
 
     expect(login.status).toBe(200);
     const token = login.body.meta.token;
-    const category = await request(app).post("/api/v1/categorias").send({
-      nombre_categoria: "Pinturas",
+    const category = await request(app).post('/api/v1/categorias').send({
+      nombre_categoria: 'Pinturas',
     });
     expect(category.status).toBe(201);
 
     const dataProduct = getProduct();
 
     const responseProduct = await request(app)
-      .post("/api/v1/productos")
-      .set("Authorization", `Bearer ${token}`)
+      .post('/api/v1/productos')
+      .set('Authorization', `Bearer ${token}`)
       .set(
-        "Content-Type",
-        `multipart/form-data; boundary=${dataProduct.getBoundary()}`
+        'Content-Type',
+        `multipart/form-data; boundary=${dataProduct.getBoundary()}`,
       )
       .send(dataProduct.getBuffer());
 
     expect(responseProduct.status).toBe(201);
 
-    const productId = "e258b309-c5db-45ea-ba0f-aa0e10f2272f";
+    const productId = 'e258b309-c5db-45ea-ba0f-aa0e10f2272f';
 
     const response = await request(app)
       .get(`/api/v1/productos/${productId}`)
-      .set("Authorization", `Bearer ${token}`);
+      .set('Authorization', `Bearer ${token}`);
     expect(response.status).toBe(404);
   }, 10000);
 
-  test("Update Product and Get my products", async () => {
+  test('Update Product and Get my products', async () => {
     await setup();
-    const login = await request(app).post("/api/v1/auth/signin").send({
-      correo: "correo_empresa@example.com",
-      contrasena: "Contra123456",
+    const login = await request(app).post('/api/v1/auth/signin').send({
+      correo: 'correo_empresa@example.com',
+      contrasena: 'Contra123456',
     });
 
     expect(login.status).toBe(200);
     const token = login.body.meta.token;
-    const category = await request(app).post("/api/v1/categorias").send({
-      nombre_categoria: "Pinturas",
+    const category = await request(app).post('/api/v1/categorias').send({
+      nombre_categoria: 'Pinturas',
     });
     expect(category.status).toBe(201);
 
     const dataProduct = getProduct();
 
     const responseProduct = await request(app)
-      .post("/api/v1/productos")
-      .set("Authorization", `Bearer ${token}`)
+      .post('/api/v1/productos')
+      .set('Authorization', `Bearer ${token}`)
       .set(
-        "Content-Type",
-        `multipart/form-data; boundary=${dataProduct.getBoundary()}`
+        'Content-Type',
+        `multipart/form-data; boundary=${dataProduct.getBoundary()}`,
       )
       .send(dataProduct.getBuffer());
     expect(responseProduct.status).toBe(201);
@@ -236,44 +237,44 @@ describe("Products Test", () => {
 
     const response = await request(app)
       .put(`/api/v1/productos/${productId}`)
-      .set("Authorization", `Bearer ${token}`)
+      .set('Authorization', `Bearer ${token}`)
       .set(
-        "Content-Type",
-        `multipart/form-data; boundary=${dataUpdate.getBoundary()}`
+        'Content-Type',
+        `multipart/form-data; boundary=${dataUpdate.getBoundary()}`,
       )
       .send(dataUpdate.getBuffer());
     expect(response.status).toBe(200);
-    expect(response.body.data.nombre).toBe("ProductoAtualizado");
+    expect(response.body.data.nombre).toBe('ProductoAtualizado');
 
     const responseMyProducts = await request(app)
-      .get("/api/v1/productos/misProductos")
-      .set("Authorization", `Bearer ${token}`);
+      .get('/api/v1/productos/misProductos')
+      .set('Authorization', `Bearer ${token}`);
 
     expect(responseMyProducts.body.data.length).toBe(1);
   }, 10000);
 
-  test("Delete Product", async () => {
+  test('Delete Product', async () => {
     await setup();
-    const login = await request(app).post("/api/v1/auth/signin").send({
-      correo: "correo_empresa@example.com",
-      contrasena: "Contra123456",
+    const login = await request(app).post('/api/v1/auth/signin').send({
+      correo: 'correo_empresa@example.com',
+      contrasena: 'Contra123456',
     });
 
     expect(login.status).toBe(200);
     const token = login.body.meta.token;
-    const category = await request(app).post("/api/v1/categorias").send({
-      nombre_categoria: "Pinturas",
+    const category = await request(app).post('/api/v1/categorias').send({
+      nombre_categoria: 'Pinturas',
     });
     expect(category.status).toBe(201);
 
     const dataProduct = getProduct();
 
     const responseProduct = await request(app)
-      .post("/api/v1/productos")
-      .set("Authorization", `Bearer ${token}`)
+      .post('/api/v1/productos')
+      .set('Authorization', `Bearer ${token}`)
       .set(
-        "Content-Type",
-        `multipart/form-data; boundary=${dataProduct.getBoundary()}`
+        'Content-Type',
+        `multipart/form-data; boundary=${dataProduct.getBoundary()}`,
       )
       .send(dataProduct.getBuffer());
     expect(responseProduct.status).toBe(201);
@@ -282,32 +283,32 @@ describe("Products Test", () => {
 
     const response = await request(app)
       .delete(`/api/v1/productos/${productId}`)
-      .set("Authorization", `Bearer ${token}`);
+      .set('Authorization', `Bearer ${token}`);
     expect(response.status).toBe(204);
   }, 10000);
 
-  test("Search Product", async () => {
+  test('Search Product', async () => {
     await setup();
-    const login = await request(app).post("/api/v1/auth/signin").send({
-      correo: "correo_empresa@example.com",
-      contrasena: "Contra123456",
+    const login = await request(app).post('/api/v1/auth/signin').send({
+      correo: 'correo_empresa@example.com',
+      contrasena: 'Contra123456',
     });
 
     expect(login.status).toBe(200);
     const token = login.body.meta.token;
-    const category = await request(app).post("/api/v1/categorias").send({
-      nombre_categoria: "Pinturas",
+    const category = await request(app).post('/api/v1/categorias').send({
+      nombre_categoria: 'Pinturas',
     });
     expect(category.status).toBe(201);
 
     const dataProduct = getProduct();
 
     const responseProduct = await request(app)
-      .post("/api/v1/productos")
-      .set("Authorization", `Bearer ${token}`)
+      .post('/api/v1/productos')
+      .set('Authorization', `Bearer ${token}`)
       .set(
-        "Content-Type",
-        `multipart/form-data; boundary=${dataProduct.getBoundary()}`
+        'Content-Type',
+        `multipart/form-data; boundary=${dataProduct.getBoundary()}`,
       )
       .send(dataProduct.getBuffer());
     expect(responseProduct.status).toBe(201);
@@ -316,120 +317,120 @@ describe("Products Test", () => {
 
     const response = await request(app)
       .get(`/api/v1/productos/search/${productName}`)
-      .set("Authorization", `Bearer ${token}`);
+      .set('Authorization', `Bearer ${token}`);
 
-    expect(response.body.data[0].nombre).toBe("ProductTest");
+    expect(response.body.data[0].nombre).toBe('ProductTest');
   }, 10000);
 
-  test("Filter Product", async () => {
+  test('Filter Product', async () => {
     await setup();
-    const login = await request(app).post("/api/v1/auth/signin").send({
-      correo: "correo_empresa@example.com",
-      contrasena: "Contra123456",
+    const login = await request(app).post('/api/v1/auth/signin').send({
+      correo: 'correo_empresa@example.com',
+      contrasena: 'Contra123456',
     });
 
     expect(login.status).toBe(200);
     const token = login.body.meta.token;
-    const category = await request(app).post("/api/v1/categorias").send({
-      nombre_categoria: "Pinturas",
+    const category = await request(app).post('/api/v1/categorias').send({
+      nombre_categoria: 'Pinturas',
     });
     expect(category.status).toBe(201);
 
     const dataProduct = getProduct();
 
     const responseProduct = await request(app)
-      .post("/api/v1/productos")
-      .set("Authorization", `Bearer ${token}`)
+      .post('/api/v1/productos')
+      .set('Authorization', `Bearer ${token}`)
       .set(
-        "Content-Type",
-        `multipart/form-data; boundary=${dataProduct.getBoundary()}`
+        'Content-Type',
+        `multipart/form-data; boundary=${dataProduct.getBoundary()}`,
       )
       .send(dataProduct.getBuffer());
     expect(responseProduct.status).toBe(201);
 
     const response = await request(app)
       .get(`/api/v1/productos/filter?filterCategoria=Pinturas`)
-      .set("Authorization", `Bearer ${token}`);
+      .set('Authorization', `Bearer ${token}`);
 
-    expect(response.body.data[0].categoria.nombre_categoria).toBe("Pinturas");
+    expect(response.body.data[0].categoria.nombre_categoria).toBe('Pinturas');
   }, 10000);
 });
 
-describe("Valorations Test", () => {
+describe('Valorations Test', () => {
   beforeEach(async () => {
     await resetDb();
   });
 
-  test("Create Valoration and GetAll", async () => {
+  test('Create Valoration and GetAll', async () => {
     await setupV();
 
-    const login = await request(app).post("/api/v1/auth/signin").send({
-      correo: "correo_ser@example.com",
-      contrasena: "Contra123456",
+    const login = await request(app).post('/api/v1/auth/signin').send({
+      correo: 'correo_ser@example.com',
+      contrasena: 'Contra123456',
     });
 
     expect(login.status).toBe(200);
     const token = login.body.meta.token;
 
-    const product = await request(app).get("/api/v1/productos/");
+    const product = await request(app).get('/api/v1/productos/');
     const idProduct = product.body.data[0].id;
 
     const responseValoration = await request(app)
       .post(`/api/v1/productos/${idProduct}/valoraciones`)
-      .set("Authorization", `Bearer ${token}`)
+      .set('Authorization', `Bearer ${token}`)
       .send({
         calificacion: 5,
-        comentarios: "Excelente producto",
+        comentarios: 'Excelente producto',
       });
 
     expect(responseValoration.status).toBe(201);
     const listValoration = await request(app).get(
-      `/api/v1/productos/${idProduct}/valoraciones`
+      `/api/v1/productos/${idProduct}/valoraciones`,
     );
     expect(listValoration.body.data.length).toBe(1);
   }, 20000);
 
-  test("Fail Create valoration for company", async () => {
+  test('Fail Create valoration for company', async () => {
     await setupV();
 
-    const login = await request(app).post("/api/v1/auth/signin").send({
-      correo: "correo_empresa@example.com",
-      contrasena: "Contra123456",
+    const login = await request(app).post('/api/v1/auth/signin').send({
+      correo: 'correo_empresa@example.com',
+      contrasena: 'Contra123456',
     });
     expect(login.status).toBe(200);
     const token = login.body.meta.token;
-    const product = await request(app).get("/api/v1/productos/");
+    const product = await request(app).get('/api/v1/productos/');
     const idProduct = product.body.data[0].id;
 
     const responseValoration = await request(app)
       .post(`/api/v1/productos/${idProduct}/valoraciones`)
-      .set("Authorization", `Bearer ${token}`)
+      .set('Authorization', `Bearer ${token}`)
       .send({
         calificacion: 5,
-        comentarios: "Excelente producto",
+        comentarios: 'Excelente producto',
       });
 
     expect(responseValoration.status).toBe(404);
   }, 20000);
 
-  test("Get ALL valorations", async () => {
+  test('Get ALL valorations', async () => {
     await setupV();
 
-    const login = await request(app).post("/api/v1/auth/signin").send({
-      correo: "correo_ser@example.com",
-      contrasena: "Contra123456",
+    const login = await request(app).post('/api/v1/auth/signin').send({
+      correo: 'correo_ser@example.com',
+      contrasena: 'Contra123456',
     });
     expect(login.status).toBe(200);
     const token = login.body.meta.token;
-    const product = await request(app).get("/api/v1/productos/");
+    const product = await request(app).get('/api/v1/productos/');
     const idProduct = product.body.data[0].id;
 
     const responseValoration = await request(app)
       .post(`/api/v1/productos/${idProduct}/valoraciones`)
-      .set("Authorization", `Bearer ${token}`)
+      .set('Authorization', `Bearer ${token}`)
       .send({
         calificacion: 5,
-        comentarios: "Excelente producto",
+        comentarios: 'Excelente producto',
       });
 
     expect(responseValoration.status).toBe(201);
